move register and signIn into user controller
